requirements.txt
# ============================================
# Project Pulse — Python Dependencies
# ============================================
# Install: pip install -r requirements.txt

# --- Orchestration ---
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0

# --- LLM Abstraction ---
langchain-google-genai>=2.0.0
langchain-core>=0.3.0

# --- Telegram Interface ---
python-telegram-bot>=21.0

# --- Database ---
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# --- Validation & Config ---
pydantic>=2.0.0
pydantic-settings>=2.0.0

# --- Utilities ---
python-dotenv>=1.0.0

# --- Dev & Testing ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
